common/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "centrality-common"
version = "0.0.1"
authors = [
    {name = "Armand McQueen", email = "armandmcqueen@gmail.com"},
]
description = "Centrality common lib"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic ~= 2.5.2",
    "PyYAML ~= 6.0",
]

[project.optional-dependencies]
dev = [
    'ruff ~= 0.1.7',
    'flit ~= 3.9.0',
    "pytest ~= 7.4.3",
    "mypy ~= 1.8.0",
    "types-PyYAML ~= 6.0.12.12",
    "types-requests ~= 2.31.0.10",
]

[tool.flit.module]
name = "common"

